Refactor the cloud compiler code into a new module. Refactor empty project message into the modal module.
