+++ src/modules/modals.js
@@ -25,6 +25,7 @@ import 'bootstrap/js/modal';
 import 'jquery-validation';
 import {displayProjectName} from './editor.js';
 import {getProjectInitialState} from './project';
+import {utils} from './utility';
 
 /**
  *  Validate the required elements of the edit project form
@@ -130,7 +131,7 @@ function setEditOfflineProjectDetailsContinueHandler() {
 function updateProjectDetails() {
   updateProjectNameDescription(
       $('#edit-project-name').val(),
-      $('#edit-project-description').val()
+      $('#edit-project-description').val(),
   );
 }
 
@@ -171,3 +172,11 @@ function setEditOfflineProjectDetailsCancelHandler() {
   });
 }
 
+/**
+ * Display a dialog window that warns of an attempt to compile
+ * an empty project.
+ */
+export const showCannotCompileEmptyProject = () => {
+  utils.showMessage(Blockly.Msg.DIALOG_EMPTY_PROJECT,
+      Blockly.Msg.DIALOG_CANNOT_COMPILE_EMPTY_PROJECT);
+};
